Parse configuration sections in configuration file.
Some times we need to group configuration keys, for example to separate client
configuration from server configuration and use the same conf file for both
situations.
All conf key that are at the beginning of the file and aren't under some section
then this keys are in automatic inserted under default section.